pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clade"
version = "0.3.0"
description = "The Clade — infrastructure for a family of Claude Code instances"
authors = [{name = "Ian Dunn", email = "ian@iandunn.io"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=0.1.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "aiosqlite>=0.20.0",
]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
clade = "clade.cli.main:cli"
clade-personal = "clade.mcp.server_full:main"
clade-worker = "clade.mcp.server_lite:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
